Refining Locations admin functionality
